Move some common fixtures to conftest.
